refactor(bootstrap): split builders into entrypoint/{daemon,cli}
Phase 4.4 of the uc-bootstrap redesign. builders.rs mixed two unrelated
entry constructors. Split by scenario: the daemon-lifecycle entry
(DaemonLifecycle + build_daemon_lifecycle) moves to entrypoint/daemon.rs,
and the CLI dev-tools wiring (build_core + build_cli_wiring_context) moves
to entrypoint/cli.rs, each carrying only the imports it needs. The
build_daemon_lifecycle/DaemonLifecycle contract re-exports source from
entrypoint::daemon; non_gui's call updated to entrypoint::cli. No behavior
change.
