+++ crates/uc-bootstrap/src/entrypoint/mod.rs
@@ -4,4 +4,6 @@
 //! daemon lifecycle, the CLI dev-tools in-process facade, and the headless
 //! (non-GUI) runtime shared by both.
 
+pub mod cli;
+pub mod daemon;
 pub mod non_gui;
